refactor(Analytics): implement useChart hook for map chart options and translations in TaskA
